requirements.txt
# Core ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Time series analysis
statsmodels>=0.13.0

# MLOps and Experiment Tracking
mlflow>=2.0.0
optuna>=3.0.0
dvc>=3.0.0
dvc-s3>=2.0.0

# API Framework
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
python-multipart>=0.0.5

# Data processing
joblib>=1.2.0
kagglehub>=0.1.0

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
typing-extensions>=4.4.0
pyyaml>=6.0

# Template engine for model cards
jinja2>=3.1.0

# Data drift detection
evidently>=0.4.12

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
isort>=5.12.0

# Jupyter environment
jupyter>=1.0.0
jupyterlab>=3.5.0
ipykernel>=6.15.0

# Monitoring and logging
prometheus-client>=0.16.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0